requirements/base.txt
Django>=1.11
django-environ>=0.4.3
django-braces>=1.11.0
django-crispy-forms>=1.6.1
django-admin-bootstrapped>=2.5.7
django-authtools>=1.4.0
easy-thumbnails>=2.4.1
psycopg2>=2.7
django-import-export>=0.5.1
django-filter>=1.0.4
python-decouple==3.0
django-redis
reportlab
django-autocomplete-light
djangorestframework>=3.7.1
markdown>=2.6.9
django-rest-auth>=0.9.2
djangorestframework-jwt>=1.11.0
